requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
pgvector
jinja2
python-dotenv
httpx
pytest
pytest-asyncio
authlib
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
